requirements.txt
pillow 
opencv-python
numpy
albumentations
matplotlib 
seaborn
tqdm
tensorboard
pyyaml
scikit-learn
torchmetrics

